consume_and_send/requirements.txt
elasticsearch<8.15.0
hashlib
pymongo
faster_whisper
kafka-python